remove getting of channels and metadata from Communities list and break out to individual methods, actions and put loop in CommunityView
