+++ src/store/index.ts
@@ -8,16 +8,23 @@ import path from "path";
 import { v4 as uuidv4 } from "uuid";
 import { createChannel } from "@/core/methods/createChannel";
 import { createProfile } from "@/core/methods/createProfile";
+import { joinChannelFromSharedLink } from "@/core/methods/joinChannelFromSharedLink";
 import { createExpression } from "@/core/mutations/createExpression";
 import { createUniqueExpressionLanguage } from "@/core/mutations/createUniqueExpressionLanguage";
 import { publishSharedPerspective } from "@/core/mutations/publishSharedPerspective";
 import { addPerspective } from "@/core/mutations/addPerspective";
 import { createLink } from "@/core/mutations/createLink";
 import { getLanguage } from "@/core/queries/getLanguage";
+import { getExpression } from "@/core/queries/getExpression";
+import {
+  getChatChannelLinks,
+  getGroupExpressionLinks,
+} from "@/core/queries/getLinks";
 import { getPerspective } from "@/core/queries/getPerspective";
 import { installSharedPerspective } from "@/core/mutations/installSharedPerspective";
 import { getTypedExpressionLanguages } from "@/core/methods/getTypedExpressionLangs";
 import sleep from "@/utils/sleep";
+import { expressionGetRetries, expressionGetDelayMs } from "@/core/juntoTypes"
 
 export interface CommunityState {
   //NOTE: here by having a static name + description we are assuming that these are top level metadata items that each group will have
@@ -444,6 +451,84 @@ export default createStore({
         typedExpressionLanguages: typedExpressionLanguages,
         sharedPerspectiveUrl: joiningLink, //TODO: this will have to be string split once we add proof onto the URL
       });
+    },
+    async getPerspectiveChannelsAndMetadata({ commit }, { community }) {
+      console.log("Getting channel links");
+      const channelLinks = await getChatChannelLinks(
+        community.perspective,
+        community.linkLanguageAddress
+      );
+      console.log("Got links", channelLinks);
+      if (channelLinks != null) {
+        for (let i = 0; i < channelLinks.length; i++) {
+          if (
+            community.channels.find(
+              (element: ChannelState) =>
+                element.sharedPerspectiveUrl === channelLinks[i].data!.target
+            ) == undefined
+          ) {
+            console.log(
+              "Found channel link",
+              channelLinks[i],
+              "Adding to channel"
+            );
+            const channel = await joinChannelFromSharedLink(
+              channelLinks[i].data!.target!
+            );
+            commit("addChannel", {
+              community: community.perspective,
+              channel: channel,
+            });
+          }
+        }
+        //NOTE/TODO: if this becomes too heavy for certain communities this might be best executed via a refresh button
+        const groupExpressionLinks = await getGroupExpressionLinks(
+          //@ts-ignore
+          community.perspective,
+          //@ts-ignore
+          community.linkLanguageAddress
+        );
+        if (groupExpressionLinks != null && groupExpressionLinks.length > 0) {
+          if (
+            community.groupExpressionRef !=
+            groupExpressionLinks[0].data!.target!
+          ) {
+            let getExprRes = await getExpression(
+              groupExpressionLinks[0].data!.target!
+            );
+            if (getExprRes == null) {
+              for (let i = 0; i < expressionGetRetries; i++) {
+                console.log("Retrying get of expression signal");
+
+                getExprRes = await getExpression(
+                  groupExpressionLinks[0].data!.target!
+                );
+                if (getExprRes != null) {
+                  break;
+                }
+                await sleep(expressionGetDelayMs);
+              }
+              if (getExprRes == null) {
+                console.warn(
+                  "Could not get expression from group expression link"
+                );
+                return;
+              }
+            }
+            const groupExpData = JSON.parse(getExprRes.data!);
+            console.log(
+              "Got new group expression data for community",
+              groupExpData
+            );
+            commit("updateCommunityMetadata", {
+              community: community.perspective,
+              name: groupExpData["foaf:name"],
+              description: groupExpData["foaf:description"],
+              groupExpressionRef: groupExpressionLinks[0].data!.target,
+            });
+          }
+        }
+      }
     },
   },
   getters: {
