+++ src/views/community-view/CommunityView.vue
@@ -5,10 +5,62 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
+import { defineComponent, onMounted, onUnmounted, ref, watch } from "vue";
+import { useRoute } from "vue-router";
 import LeftDrawer from "../../components/left-drawer/LeftDrawer.vue";
 import MainView from "./main-view/MainView.vue";
-export default {
+import { channelRefreshDurationMs } from "@/core/juntoTypes";
+import { useStore } from "vuex";
+
+export default defineComponent({
+  setup() {
+    const route = useRoute();
+    const store = useStore();
+
+    let noDelayRef: any = ref();
+
+    watch(
+      () => route.params.communityId,
+      (params: any) => {
+        console.log("Got update", params);
+        startLoop(params);
+      }
+    );
+
+    //Dont think we need this
+    // onMounted(() => {
+    //   let communityId = route.params.communityId;
+    //   if (communityId != undefined) {
+    //     startLoop(communityId);
+    //   }
+    // });
+
+    onUnmounted(() => {
+      clearInterval(noDelayRef.value);
+    });
+
+    function startLoop(communityId: string) {
+      clearInterval(noDelayRef.value);
+      if (communityId != null) {
+        console.log("Running get channels loop");
+        const test = noDelaySetInterval(async () => {
+          store.dispatch("getPerspectiveChannelsAndMetadata", {
+            community: store.getters.getCommunity(communityId),
+          });
+        }, channelRefreshDurationMs);
+
+        noDelayRef.value = test;
+      }
+    }
+
+    //TODO: @leif idk the best place to put this
+    function noDelaySetInterval(func: () => void, interval: number) {
+      func();
+
+      return setInterval(func, interval);
+    }
+  },
   components: {
     LeftDrawer,
     MainView,
@@ -19,7 +71,7 @@ export default {
       return this.$store.getters.getCommunity(communityId);
     },
   },
-};
+});
 </script>
 
 <style lang="scss" scoped>
